# GH-3416 Cleanup and further optimization of LMDB store.