--- a/core/sail/lmdb/src/main/java/org/eclipse/rdf4j/sail/lmdb/ValueStoreRevision.java
+++ b/core/sail/lmdb/src/main/java/org/eclipse/rdf4j/sail/lmdb/ValueStoreRevision.java
@@ -47,4 +47,8 @@ public class ValueStoreRevision implements Serializable {
 	public ValueStore getValueStore() {
 		return valueStore;
 	}
+
+	public boolean resolveValue(long id, LmdbValue value) {
+		return valueStore.resolveValue(id, value);
+	}
 }