@kenwenzel 2021-12-13
Before:
```
Benchmark                                               Mode  Cnt     Score     Error  Units
QueryBenchmark.distinctPredicatesQuery                  avgt    5  1231.341 ± 120.133  ms/o
```

After:
```
Benchmark                               Mode  Cnt    Score    Error  Units
QueryBenchmark.distinctPredicatesQuery  avgt    5  525.137 ± 24.159  ms/op
```