@kenwenzel 2021-12-14
Before simple value cache:
```
Benchmark                                               Mode  Cnt     Score     Error  Units
QueryBenchmark.complexQuery                             avgt    5    31.438 ±   6.007  ms/op
QueryBenchmark.distinctPredicatesQuery                  avgt    5   534.606 ±  53.634  ms/op
QueryBenchmark.groupByQuery                             avgt    5    11.344 ±   1.461  ms/op
```

With simple value cache:
```
Benchmark                               Mode  Cnt    Score    Error  Units
QueryBenchmark.complexQuery             avgt    5   29.841 ±  1.027  ms/op
QueryBenchmark.distinctPredicatesQuery  avgt    5  425.238 ± 18.046  ms/op
QueryBenchmark.groupByQuery             avgt    5    6.783 ±  0.260  ms/op
```