@JervenBolleman 2021-12-15
I love the performance improvement. Really nice, I have a few nits around locking. I also want to double check the serialization code, I am not very familiar with it but it currently seems off to me, but I might well be wrong about it :)