@kenwenzel 2021-12-16
> I love the performance improvement. Really nice, I have a few nits around locking. I also want to double check the serialization code, I am not very familiar with it but it currently seems off to me, but I might well be wrong about it :)

The `writeReplace` methods ensure that only resolved values are serialized so that the values are fully initialized after deserialization.
Locking of the array used for caching should not be required. The value ID is checked after reading and when new values are inserted only one field of the array is overwritten.