@kenwenzel 2021-12-17
```
# lmdb-lazy BEFORE
Benchmark                                               Mode  Cnt     Score     Error  Units
QueryBenchmark.complexQuery                             avgt    5    29.223 ±   4.899  ms/op
QueryBenchmark.distinctPredicatesQuery                  avgt    5   454.409 ±  62.850  ms/op
QueryBenchmark.groupByQuery                             avgt    5     6.847 ±   0.668  ms/op
QueryBenchmark.removeByQuery                            avgt    5   352.492 ± 284.050  ms/op
QueryBenchmark.removeByQueryReadCommitted               avgt    5   765.650 ± 180.301  ms/op
QueryBenchmark.simpleUpdateQueryIsolationNone           avgt    5   484.997 ±  58.529  ms/op
QueryBenchmark.simpleUpdateQueryIsolationReadCommitted  avgt    5  1172.697 ±  87.222  ms/op
```

```
# lmdb-lazy NOW
Benchmark                                               Mode  Cnt     Score     Error  Units
QueryBenchmark.complexQuery                             avgt    5   22.968 ± 1.477   ms/op
QueryBenchmark.distinctPredicatesQuery                  avgt    5  358.696 ± 25.942  ms/op
QueryBenchmark.groupByQuery                             avgt    5    6.057 ±  0.236  ms/op
QueryBenchmark.removeByQuery                            avgt    5  241.248 ± 14.692  ms/op
QueryBenchmark.removeByQueryReadCommitted               avgt    5  659.642 ± 35.941  ms/op
QueryBenchmark.simpleUpdateQueryIsolationNone           avgt    5  413.846 ± 15.673  ms/op
QueryBenchmark.simpleUpdateQueryIsolationReadCommitted  avgt    5  987.459 ± 90.607  ms/op
```