@jeenbroekstra 2022-01-07
> Can this be merged to develop so that I can base further work on it?

Of course! Sorry, I fogot that you aren't a committer for a sec :) 

By the way: I am tentatively aiming for a new milestone build on January 22 - does that give you enough time to get the LMDB store in the state you'd want it to be? This won't be an official release yet, but we put it out for early adopter feedback.